ref(spans): Rename detect-performance-issues to process-segments
