+++ src/sentry/options/defaults.py
@@ -2650,7 +2650,7 @@ register(
     flags=FLAG_PRIORITIZE_DISK | FLAG_AUTOMATOR_MODIFIABLE,
 )
 register(
-    "standalone-spans.detect-performance-issues-consumer.enable",
+    "standalone-spans.process-segments.enable",
     default=True,
     flags=FLAG_PRIORITIZE_DISK | FLAG_AUTOMATOR_MODIFIABLE,
 )
