+++ src/sentry/runner/commands/devserver.py
@@ -357,7 +357,7 @@ def devserver(
             if settings.SENTRY_USE_SPANS_BUFFER:
                 kafka_consumers.add("process-spans")
                 kafka_consumers.add("ingest-occurrences")
-                kafka_consumers.add("detect-performance-issues")
+                kafka_consumers.add("process-segments")
 
         if occurrence_ingest:
             kafka_consumers.add("ingest-occurrences")
