+++ src/sentry/utils/sdk.py
@@ -39,7 +39,7 @@ logger = logging.getLogger(__name__)
 UNSAFE_FILES = (
     "sentry/event_manager.py",
     "sentry/spans/consumers/process/factory.py",
-    "sentry/spans/consumers/detect_performance_issues/factory.py",
+    "sentry/spans/consumers/process_segments/factory.py",
     "sentry/tasks/process_buffer.py",
     "sentry/ingest/consumer/processors.py",
     # This consumer lives outside of sentry but is just as unsafe.
