+++ tests/sentry/consumers/test_run.py
@@ -35,7 +35,7 @@ def test_dlq(consumer_def) -> None:
     ]
     consumers_that_should_have_dlq_but_dont = [
         "process-spans",
-        "detect-performance-issues",
+        "process-segments",
         "ingest-monitors",
         "monitors-clock-tick",
         "monitors-clock-tasks",
